+++ Frends.Pgp.EncryptFile/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.1.0] - 2026-07-16
+
+### Added
+
+- Support public keys provided directly as a string during encryption
+- Support signing private keys provided directly as a string, including Utf8 and Legacy passphrase handling
+
 ## [1.0.0] - 2025-12-12
 
 ### Added